Revert "fix(agent-hooks): bind agent status to the pane its session was spawned into (STA-2069) (#14615)"
This reverts commit 15efc87e35c5e00b3ba3922ef9e9bc55769599a0.

The premise does not reproduce, the fix does not reach the failure that does,
and the correction it installs can misattribute status on a path that worked
before.

1. PREMISE FALSE. #14615 asserts Claude Code >= 2.1.206 hosts TUI sessions as
   workers under a shared daemon. On 2.1.233 `claude daemon status` reports
   "not running" while 69 interactive sessions are live, and all 68 running
   clients are direct children of their own pane's shell. Measured across the
   fleet: 68 distinct pane keys, ZERO collisions. The 19 keys with no live pane
   also have a dead ORCA_TERMINAL_HANDLE (pane closed, process outlived it).
   Terminals live under a *different* pane key: 0. Foreground attribution was
   never broken, so the correction has nothing to correct.

2. DOES NOT FIX THE REAL BUG. The failure #9236 describes is real but scoped to
   BACKGROUNDED sessions: a `--bg` / `/background` worker inherits the
   dispatching pane's whole ORCA_* set, and the hook script self-heals its
   endpoint while taking paneKey straight from ambient env, so the POST lands
   under the wrong pane. #14615 mints a binding only for launches Orca
   constructs, so a typed `claude --bg` produces none. That needs a `kind: "bg"`
   guard at ingest instead; #9236 is reopened for it.

3. INTRODUCES A MISATTRIBUTION. Bindings are created at two spawn sites and
   removed only on PTY death. A pane whose user exits Claude keeps its PTY, so
   the binding survives; resuming that session in another pane does not rebind
   (`--resume` is a session selector, so the pin declines), and
   resolveBoundPaneOverride then rewrites paneKey and tabId onto the ORIGINAL
   pane despite a correct posted key. That is the reported symptom, on a path
   that previously worked.

Kept #14706's observations.rebind() in the conflicting hunk — it postdates
#14615 and is not part of this revert; only #14615's comments were dropped and
the surviving rationale reworded to alias resolution, which is what actually
makes paneKey differ from envelope.paneKey now.
