+++ src/main/ipc/pty-ipc-suite-environment.ts
@@ -32,8 +32,6 @@ import {
   setMigrationUnsupportedPtyMock,
   clearMigrationUnsupportedPtyMock,
   clearMigrationUnsupportedPtysForPaneKeyMock,
-  bindAgentSessionPaneMock,
-  clearAgentSessionPaneBindingsForPtyMock,
   clearPaneKeyAliasesForPtyMock,
   recordCodexPaneAccountMock,
   forgetCodexPaneAccountMock,
@@ -125,8 +123,6 @@ export function createPtyIpcSuiteEnvironment(): PtyIpcSuiteEnvironment {
     setMigrationUnsupportedPtyMock.mockReset()
     clearMigrationUnsupportedPtyMock.mockReset()
     clearMigrationUnsupportedPtysForPaneKeyMock.mockReset()
-    bindAgentSessionPaneMock.mockReset()
-    clearAgentSessionPaneBindingsForPtyMock.mockReset()
     clearPaneKeyAliasesForPtyMock.mockReset()
     recordCodexPaneAccountMock.mockReset()
     forgetCodexPaneAccountMock.mockReset()
