+++ src/main/ipc/pty.ts
@@ -118,9 +118,6 @@ import {
   markClaudePtyExited,
   markClaudePtySpawned
 } from '../claude-accounts/live-pty-gate'
-import { pinClaudeLaunchSessionId } from '../../shared/claude-session-pin-launch-command'
-import type { AgentStartupShell } from '../../shared/tui-agent-startup-shell'
-import { resolveSpawnStartupShell } from '../pty/spawn-startup-shell'
 import { ensureLinuxTerminalOrcaCliShimDir } from '../cli/linux-terminal-orca-cli-shim'
 import {
   isLegacyTerminalShimPathEntry,
@@ -2007,31 +2004,6 @@ function isClaudeLaunchCommand(command: string | undefined): boolean {
   )
 }
 
-type ClaudeSessionPinPlan = { command: string | undefined; sessionId: string | null }
-
-/** Mints a session id into a Claude launch so hook events can be traced back to
- * the pane Orca spawned them into.
- *
- * Why: Claude Code >= 2.1.206 hosts TUI sessions as workers under a shared
- * daemon, and the daemon forwards only its own allowlisted env — so a worker's
- * `ORCA_PANE_KEY` names whichever pane first started the daemon, not the pane
- * the user is looking at (#9236). The session id is the only pane-independent
- * identity a hook payload carries, so binding it here is what lets the hook
- * server route status to the true pane. Pinning is best-effort by design: a
- * command it cannot splice safely is left untouched and keeps today's behavior.
- */
-function planClaudeSessionPin(
-  command: string | undefined,
-  shell: AgentStartupShell
-): ClaudeSessionPinPlan {
-  if (command === undefined || !isClaudeLaunchCommand(command)) {
-    return { command, sessionId: null }
-  }
-  const sessionId = randomUUID()
-  const pinned = pinClaudeLaunchSessionId(command, sessionId, shell)
-  return pinned ? { command: pinned, sessionId } : { command, sessionId: null }
-}
-
 function routesFreshSpawnsToLocalProvider(provider: IPtyProvider): boolean {
   return provider.routesFreshSpawnsToLocalProvider === true
 }
@@ -2159,9 +2131,6 @@ export function clearProviderPtyState(
   piTitlebarExtensionService.clearPty(id)
   // Why: SSH exit/teardown paths bypass pty.ts's local onExit but still must release Claude account-switch guards.
   markClaudePtyExited(id)
-  // Why: a dead PTY's session pin must not re-route a later session onto a pane
-  // it no longer runs.
-  agentHookServer.clearAgentSessionPaneBindingsForPty(id)
   ptySizes.delete(id)
   ptyIncarnationById.delete(id)
   lastInputAtByPty.delete(id)
@@ -4692,18 +4661,7 @@ export function registerPtyHandlers(
       const codexResumeHome = codexResumeLaunch.codexResumeHome
       // Why: the drop still applies here, but this controller's result has no field for
       // notifyResumeUnavailable — runtime/relay panes start fresh without the notice.
-      // Why not gated on isClaudeLaunch: that flag is local-only (it guards
-      // account switching); a remote host's daemon inherits a stale pane key too.
-      const claudeSessionPin = planClaudeSessionPin(
-        codexResumeLaunch.command,
-        resolveSpawnStartupShell({
-          connectionId: args.connectionId,
-          windowsWslDistro: terminalRuntimeOptions.terminalWindowsWslDistro,
-          shellOverride: daemonShellOverride,
-          platform: process.platform
-        })
-      )
-      const launchCommand = claudeSessionPin.command
+      const launchCommand = codexResumeLaunch.command
       const claudeAuth =
         isClaudeLaunch && prepareClaudeAuth ? await prepareClaudeAuth(codexSelectionTarget) : null
       if (isClaudeLaunch && isClaudeAuthSwitchInProgress()) {
@@ -5473,13 +5431,6 @@ export function registerPtyHandlers(
         }
         // Why: runtime-owned CLI PTYs bypass the renderer pty:spawn handler; record paneKey here too since hook titles and cache cleanup need this reverse lookup.
         const paneKey = rememberPaneKeyForPty(result.id, env?.ORCA_PANE_KEY)
-        if (claudeSessionPin.sessionId && paneKey) {
-          agentHookServer.bindAgentSessionPane('claude', claudeSessionPin.sessionId, {
-            paneKey,
-            ptyId: result.id,
-            worktreeId: args.worktreeId
-          })
-        }
         const pendingSerializer = paneKey ? pendingByPaneKey.get(paneKey) : undefined
         const inheritRendererReadiness =
           result.isReattach === true &&
@@ -6496,18 +6447,7 @@ export function registerPtyHandlers(
           ? await resolveCodexResumeLaunch(args.command, codexResumePreparation)
           : noCodexResumeLaunch(preAdoptedStablePane ? undefined : args.command)
         const codexResumeHome = codexResumeLaunch.codexResumeHome
-        // Why not gated on isClaudeLaunch: that flag is local-only (it guards
-        // account switching); a remote host's daemon inherits a stale pane key too.
-        const claudeSessionPin = planClaudeSessionPin(
-          codexResumeLaunch.command,
-          resolveSpawnStartupShell({
-            connectionId: args.connectionId,
-            windowsWslDistro: terminalRuntimeOptions.terminalWindowsWslDistro,
-            shellOverride: initialShellOverride,
-            platform: process.platform
-          })
-        )
-        const launchCommand = claudeSessionPin.command
+        const launchCommand = codexResumeLaunch.command
         baseEnv = stripSequencedStartupResumeArgv(baseEnv, codexResumeLaunch)
         // Why: declared after the strip so a local-provider spawn cannot capture the
         // pre-strip env — only the daemon branch below re-derives this from baseEnv.
@@ -7159,13 +7099,6 @@ export function registerPtyHandlers(
         const rememberedPaneKey = validatedPaneKey
           ? rememberPaneKeyForPty(result.id, validatedPaneKey)
           : null
-        if (claudeSessionPin.sessionId && rememberedPaneKey) {
-          agentHookServer.bindAgentSessionPane('claude', claudeSessionPin.sessionId, {
-            paneKey: rememberedPaneKey,
-            ptyId: result.id,
-            worktreeId: args.worktreeId
-          })
-        }
         if (legacySpawnPaneKey && migrationUnsupportedPaneKey) {
           agentHookServer.registerPaneKeyAlias(
             legacySpawnPaneKey.paneKey,
