+++ src/shared/agent-hook-listener.ts
@@ -4351,44 +4351,6 @@ function readStringField(record: Record<string, unknown>, key: string): string |
   return trimmed.length > 0 ? trimmed : undefined
 }
 
-/**
- * The provider session id a raw hook body reports, read exactly as
- * `normalizeHookPayload` reads it (same JSON guard, same per-source extractor,
- * same Codex child carve-out).
- *
- * Why it is exposed: pane attribution has to run BEFORE normalization.
- * `normalizeHookPayload` keys its prompt/compaction state machine on the posted
- * pane key, so correcting the key afterwards would leave that state on the
- * wrong pane. Callers that re-derive the session id themselves drift from this
- * parser the moment a source changes shape.
- */
-export function readHookBodyProviderSessionId(
-  source: AgentHookSource,
-  body: unknown
-): string | null {
-  if (typeof body !== 'object' || body === null) {
-    return null
-  }
-  const rawPayload = (body as Record<string, unknown>).payload
-  let hookPayload: unknown = rawPayload
-  if (typeof rawPayload === 'string') {
-    try {
-      hookPayload = parseAgentHookJson(rawPayload)
-    } catch {
-      return null
-    }
-  }
-  if (typeof hookPayload !== 'object' || hookPayload === null) {
-    return null
-  }
-  const record = hookPayload as Record<string, unknown>
-  // Why: Codex child hooks expose the child's session_id on the parent's pane.
-  if (source === 'codex' && readString(record, 'agent_id')) {
-    return null
-  }
-  return extractAgentProviderSession(source, record)?.id ?? null
-}
-
 export function normalizeHookPayload(
   state: HookListenerState,
   source: AgentHookSource,
